% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm}
\alias{cm}
\title{Matrix showing complicated management of China's Water Resource}
\format{
A data frame with 13 rows and 11 variables:
\describe{
  ...
}
}
\usage{
cm
}
\description{
Matrix showing complicated management of China's Water Resource
}
\keyword{datasets}
