% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBS.R
\name{dataJson2df}
\alias{dataJson2df}
\title{private function to convert the returned jason data to a dataframe}
\usage{
dataJson2df(rawObj, rowcode, colcode)
}
\arguments{
\item{rawObj}{the fromJSON output}

\item{rowcode}{rowcode in the data frame}

\item{colcode}{colcode in the data frame}
}
\value{
the contructed data frame
}
\description{
private function to convert the returned jason data to a dataframe
}
\references{
Xuehui YANG (2016). rstatscn: R Interface for China National Data. R
package version 1.1.1. https://CRAN.R-project.org/package=rstatscn
}
