% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSRRating.R
\name{doCSRCorp}
\alias{doCSRCorp}
\title{get CSR rating from a website}
\usage{
doCSRCorp(startPage, endPage)
}
\arguments{
\item{startPage}{on Which page you want to start, default is 1}

\item{endPage}{On which page you want to stop scrapping}
}
\value{
A table of CSR ratings collected from your input page
}
\description{
get CSR rating from a website
}

