% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDPmix}
\alias{GDPmix}
\title{Table about GDP mix of China provinces in 2015}
\format{A data frame with 11 rows and 7 variables:
\describe{
  \item{Province}{province}
  \item{x}{x axis for the plot, usually wastewater discharge per 10000 GDP}
  \item{y}{y axis for the plot, usually water use per 10000 GDP}
  \item{GDP}{China GDP 2015}
  \item{First}{GDP of first industry}
  \item{Second}{GDP of second industry}
  \item{Third}{GDP of third industry}
  ...
}}
\usage{
GDPmix
}
\description{
Table about GDP mix of China provinces in 2015
}
\keyword{datasets}
