% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSRRating.R
\name{getCSRRating}
\alias{getCSRRating}
\title{get CSR rating from a website}
\usage{
getCSRRating(startPage, endPage, date = "2015-12-31", proxy = NULL)
}
\arguments{
\item{startPage}{on Which page you want to start, default is 1}

\item{endPage}{On which page you want to stop scrapping}

\item{date}{represents the date is until which date, usually it's the last day of a year
e.g., "2015-12-31" for the date of year 2015, "2014-12-31" for the date of year 2014}

\item{proxy}{the proxy, default is NULL}
}
\value{
A table of CSR ratings collected from your input page
}
\description{
get CSR rating from a website
}

