% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSRRating.R
\name{getCSRRating_unit}
\alias{getCSRRating_unit}
\title{get CSR rating from a website for a unit page}
\usage{
getCSRRating_unit(page, date, proxy = NULL)
}
\arguments{
\item{page}{on Which page you want to scrap}

\item{date}{represents the date is until which date, usually it's the last day of a year
e.g., "2015-12-31" for the date of year 2015, "2014-12-31" for the date of year 2014}

\item{proxy}{whether use the proxy, default is FALSE}
}
\value{
A table of CSR ratings collected from your input page
}
\description{
get CSR rating from a website for a unit page
}
\details{
Get CSR ratings and reports of different companies from http://stockdata.stock.hexun.com/zrbg/
}
\references{
www.hexun.com
}
