% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getENNames.R
\name{getENNames}
\alias{getENNames}
\title{get a company's EN names}
\usage{
getENNames(tickers)
}
\arguments{
\item{tickers}{ticker/sympol of a company, TICKERS MUST BE CHARACTERs, '006027' INSTEAD OF '6027'}
}
\value{
A data table with companies' EN names
}
\description{
get a company's EN names
}
\details{
Data comes from hexun.com
}
\examples{
\dontrun{
getENNames(601857)
}


}
\references{
http://hexun.com
}
