% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getENNames.R
\name{getENNames_unit}
\alias{getENNames_unit}
\title{get a company's English name}
\usage{
getENNames_unit(ticker)
}
\arguments{
\item{ticker}{ticker/sympol of a company, MUST BE A CHARACTER, '006027' INSTEAD OF '6027'}
}
\value{
A data table with companies' EN names
}
\description{
get a company's English name
}
\details{
Data comes from hexun.com
}
\references{
http://hexun.com
}
