% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExchange.R
\name{getExchange}
\alias{getExchange}
\title{get a company's listed location}
\usage{
getExchange(tickers)
}
\arguments{
\item{tickers}{ticker/sympol of a company, TICKERS MUST BE CHARACTERs, '006027' INSTEAD OF '6027'}
}
\value{
A data table with a listed companies' ticker, security name and listed exchange location
}
\description{
get a company's listed location
}
\details{
Data comes from www.finance.sina.com.cn
}
\examples{
\dontrun{
getExchange('600601')
getExchange(c('00005', '00001'))
}
}
\references{
www.finance.sina.com.cn
}
