% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMktCap.R
\name{getHisMktCap_unit}
\alias{getHisMktCap_unit}
\title{get a company's historical market cap, data comes from NetEase}
\usage{
getHisMktCap_unit(ticker, date1, date2)
}
\arguments{
\item{ticker}{ticker/sympol of a company, MUST BE A CHARACTER, '006027' INSTEAD OF '6027'}

\item{date1}{starting date, in the following format "20160101", means Jan 1st of 2016}

\item{date2}{ending date, in the following format "20160101", if you only want one day's data, just set starting date and ending date the same day}
}
\value{
A data table with companies total capitalization and market capitalization
}
\description{
get a company's historical market cap, data comes from NetEase
}
\details{
Data comes from www.money.163.com
}
\references{
www.money.163.com
}
