% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndex}
\alias{getIndex}
\title{get a company's market cap, data comes from NetEase}
\usage{
getIndex(tickers)
}
\arguments{
\item{tickers}{ticker/sympol of a company}
}
\value{
A data table with companies and which index they are included
}
\description{
get a company's market cap, data comes from NetEase
}
\examples{
\dontrun{
getIndex(600601)
}


}
