% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndexConstnt}
\alias{getIndexConstnt}
\title{get a company's market cap, data comes from NetEase}
\usage{
getIndexConstnt(indexPool)
}
\arguments{
\item{indexPool}{a pool of different index, special format for gfer}
}
\value{
A data table with companies total capitalization and market capitalization
}
\description{
get a company's market cap, data comes from NetEase
}
