% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndexData}
\alias{getIndexData}
\title{get index information
Currently include CSI 100, SSE 50, CSI 300, SSE Central SOEs 50, HSI, HSCEI}
\usage{
getIndexData()
}
\value{
a data table containing index information
}
\description{
get index information
Currently include CSI 100, SSE 50, CSI 300, SSE Central SOEs 50, HSI, HSCEI
}
