% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMktCap.R
\name{getMktCap}
\alias{getMktCap}
\title{get a company's market cap, data comes from NetEase}
\usage{
getMktCap(tickers, date1, date2)
}
\arguments{
\item{tickers}{ticker/sympol of a company}

\item{date1}{starting date, in the following format "20160101", means Jan 1st of 2016}

\item{date2}{ending date, in the following format "20160101", if you only want one day's data, just set starting date and ending date the same day}
}
\value{
A data table with companies total capitalization and market capitalization
}
\description{
get a company's market cap, data comes from NetEase
}
\examples{
\dontrun{
getMktCap(600601, '20160101', '20160102')
}


}
