% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBS.R
\name{getNBS}
\alias{getNBS}
\title{getNBS}
\usage{
getNBS(indicator, start, end)
}
\arguments{
\item{indicator}{of which data is fetched, indicator includes 'GDP', 'water resources', 'water use' and 'wastewater', etc.}

\item{start}{starting year of data wanted}

\item{end}{end year of data wanted, make sure your input end year exists in the NBS website}
}
\value{
no return
}
\description{
get National Bureau of Statistics data
}
\references{
Xuehui YANG (2016). rstatscn: R Interface for China National Data. R
package version 1.1.1. https://CRAN.R-project.org/package=rstatscn
}
