% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProxy.R
\name{getProxy}
\alias{getProxy}
\title{Get proxy pool from free proxy provider}
\usage{
getProxy()
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Get proxy pool from free proxy provider
}
\details{
Extract proxies from http://www.free-proxy-list.net/, in case of the risk of being blocked by the scrapped website
}
\references{
www.free-proxy-list.net
}
