% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listedInfo.R
\name{getStockList}
\alias{getStockList}
\title{Get information from Shanghai Exchange and Shenzhen Exchange.
This will only get stock information in Shanghai Exchange and Shenzhen Exchange
Including stocker ticker, stock name and company full name. Data comes from China Merchants Bank}
\usage{
getStockList()
}
\description{
Get information from Shanghai Exchange and Shenzhen Exchange.
This will only get stock information in Shanghai Exchange and Shenzhen Exchange
Including stocker ticker, stock name and company full name. Data comes from China Merchants Bank
}
\references{
http://info.cmbchina.com/Stock/Single/
}
