% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTickers.R
\name{getTickers}
\alias{getTickers}
\title{get ticker by input a company's full name or a list of companies' full name}
\usage{
getTickers(corpNames)
}
\arguments{
\item{corpNames}{Full name of a company, should be full name}
}
\value{
A data table with companies stock name and stock ticker
}
\description{
It can also be a way to test if a company is listed
NOTE: If a company is listed in multiple exchange, then it needs double check,
the programe only chooses ticker from random exchange
}
\details{
Data comes from www.cninfo.com.cn/
}
\references{
www.cninfo.com.cn
}
