% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTickers.R
\name{getTickers_unit}
\alias{getTickers_unit}
\title{get ticker by input a company's full name}
\usage{
getTickers_unit(corpName)
}
\arguments{
\item{corpName}{Full name of a company}
}
\value{
A data table with companies stock name and stock ticker
}
\description{
It can also be a way to test if a company is listed
}
\details{
Data comes from www.cninfo.com.cn/
}
