% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBS.R
\name{getWaternomicsData_goog}
\alias{getWaternomicsData_goog}
\title{getWaternomicsData_goog}
\usage{
getWaternomicsData_goog()
}
\description{
Get NBS data from google sheet by shared link. Default link is provided by gfer, you can also create your
own google sheet of GDP.
NOTE: The 'link sharing on' of the sheet must be ticked in order to read
}
