% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBS.R
\name{getWaternomicsData_NBS}
\alias{getWaternomicsData_NBS}
\title{getWaternomicsData_NBS}
\usage{
getWaternomicsData_NBS(start, end)
}
\arguments{
\item{start}{starting year of data wanted}

\item{end}{end year of data wanted, make sure your input end year exists in the NBS website}
}
\description{
Get NBS data from NBS website.
}
