% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWQ.R
\name{getWaterQ_MEP_all}
\alias{getWaterQ_MEP_all}
\title{get PPP list from a single page}
\usage{
getWaterQ_MEP_all(year, week, station1, station2)
}
\arguments{
\item{year}{In which year you would like to scrape}

\item{week}{In which week you would like to scrape}

\item{station1}{the start station index on the page}

\item{station2}{the end station index on the page}
}
\description{
get PPP list from a single page
}
\details{
Get monitoring data of different stations from Minitsry of Environmental Protection of China. Using this function
you will get data of all the stations. Since the number of stations vary with time, using this function, you have
to make sure that within the period you are scrapping, the number of stations keep consistant.
}

