% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWQ.R
\name{getWaterQ_MEP_all_unit}
\alias{getWaterQ_MEP_all_unit}
\title{get PPP list from a single page}
\usage{
getWaterQ_MEP_all_unit(year, week, station1, station2, proxy = NULL)
}
\arguments{
\item{year}{In which year you would like to scrape}

\item{week}{In which week you would like to scrape}

\item{station1}{the start station index on the page}

\item{station2}{the end station index on the page}

\item{proxy}{if you wnat to use a proxy to avoid blocking, you can input a proxy, otherwise leave
it blank.}
}
\value{
A table of PPP projects collected from your input page
}
\description{
get PPP list from a single page
}
\references{
http://datacenter.mee.gov.cn/report/getCountGraph.do?type=runQianWater
}
