% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listedInfo.R
\name{is.listed}
\alias{is.listed}
\title{Check if a company is listed in Chinese stock market}
\usage{
is.listed(corpList, stockData)
}
\arguments{
\item{corpList}{company list you want to check if listed}

\item{stockData}{Result from \code{\link{getStockList}}}
}
\description{
Check if a company is listed in Chinese stock market
}
\references{
http://info.cmbchina.com/Stock/Single/
}
