% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot}
\alias{pivot}
\title{pivot}
\usage{
pivot(data, reserve, newColName)
}
\arguments{
\item{data}{a dataframe like \code{GDPmix}}

\item{reserve}{choose which column to reserve}

\item{newColName}{name for new column, apart from the reserved columns, all the other will be re-organised into a new column}
}
\description{
Transfer 2D table to readable table for tableau
}
