% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaternomics.R
\name{plotWaternomics}
\alias{plotWaternomics}
\title{plotWaternomics}
\usage{
plotWaternomics(data, title = "", xname = "", yname = "", small = 1.5,
  medium = 3, large = 5, legend = TRUE, label = TRUE)
}
\arguments{
\item{data}{a dataframe containing columns of GDP, Value-added of fisrt/second/third industry, x and y
x and y are the coordinates of the plot, and it varies depending on needs, can be per capita
water use vs. per capita wastewater, etc. See example of year 2015 by \code{GDPmix}

Columns should be exactly the same as GDPmix, which means, first column is province, seconnd is x,
third is y, fourth is Frist, etc. Load GDPmix to have a look at the columns.}

\item{title}{chart title}

\item{xname}{x axis name}

\item{yname}{y axis name}

\item{small}{radius of small circle, default is 1.5}

\item{medium}{radius of medium circle, default is 3}

\item{large}{radius of large circle, default is 5}

\item{legend}{whether to show legend, default is TRUE}

\item{label}{whether to show label, default is TRUE}
}
\description{
plot special waternomics chart for CWR
}
\examples{
\dontrun{
data(GDPmix)
plotwaternomics(GDPmix)
}
}
