% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaternomics.R
\name{plotWaternomics_legend}
\alias{plotWaternomics_legend}
\title{plotWaternomics_legend}
\usage{
plotWaternomics_legend(label = TRUE)
}
\arguments{
\item{label}{whether to show label, default is TRUE}
}
\description{
legend for plot special waternomics chart for CWR
}
\examples{
\dontrun{
plotwaternomics_legend()
}
}
