% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBS.R
\name{statscnDbs}
\alias{statscnDbs}
\title{the available dbs}
\usage{
statscnDbs()
}
\value{
a data frame with 2 columns , one is the dbcode, another is the db description
}
\description{
the available dbs in the national db
}
\examples{
 \dontrun{
 statscnDbs()
 }

}
\references{
Xuehui YANG (2016). rstatscn: R Interface for China National Data. R
package version 1.1.1. https://CRAN.R-project.org/package=rstatscn
}
