% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBS.R
\name{updateNBS}
\alias{updateNBS}
\title{updateNBS}
\usage{
updateNBS(start, end)
}
\arguments{
\item{start}{starting year of data wanted}

\item{end}{end year of data wanted, make sure your input end year exists in the NBS website}
}
\description{
update/create the database in your google sheet. You have to sign in mannually for your google sheet
Once finished, there will be a google sheet called NBS_data created in your google drive as database.
}
