% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDPmix}
\alias{GDPmix}
\title{Table about GDP mix of China provinces in 2015}
\format{
A data frame with 11 rows and 7 variables:
\describe{
  ...
}
}
\usage{
GDPmix
}
\description{
Table about GDP mix of China provinces in 2015
}
\keyword{datasets}
