% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSRRating.R
\name{getCSRRating}
\alias{getCSRRating}
\title{get CSR rating from a website}
\usage{
getCSRRating(startPage, endPage, year = 2015, proxy = FALSE)
}
\arguments{
\item{startPage}{on Which page you want to start, default is 1}

\item{endPage}{On which page you want to stop scrapping}

\item{year}{In which year you want the rank}

\item{proxy}{whether use the proxy, default is FALSE}
}
\value{
A table of CSR ratings collected from your input page
}
\description{
get CSR rating from a website
}
\details{
Get CSR ratings and reports of different companies from http://stockdata.stock.hexun.com/zrbg/
}
\examples{
\dontrun{
# get first two pages of CSR ratings in 2015
getCSRRating(1,3)
}
}
\references{
www.hexun.com
}
