% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndex}
\alias{getIndex}
\title{get a company's market cap, data comes from NetEase}
\usage{
getIndex(tickers, indexData)
}
\arguments{
\item{tickers}{ticker/sympol of a company, MUST BE A CHARACTER, e.g., input "006600" instead of 006600}

\item{indexData}{the index information, before running getIndex, indexData needs to be loaded using \code{\link{getIndexData}}}
}
\value{
A data table with companies and which index they are included
}
\description{
get a company's market cap, data comes from NetEase
}
\details{
Data comes from www.finance.sina.com.cn and www.etnet.com.hk
}
\examples{
\dontrun{
getIndex(600601)
}
}
\references{
www.finance.sina.com.cn
www.etnet.com.hk
}
