% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPPList.R
\name{getPPPList}
\alias{getPPPList}
\title{get PPP list from an official website}
\usage{
getPPPList(startPage = 1, endPage, proxy = FALSE)
}
\arguments{
\item{startPage}{on Which page you want to start, default is 1}

\item{endPage}{On which page you want to stop scrapping}

\item{proxy}{whether proxy will be used, default is FALSE}
}
\value{
A table of PPP projects collected from your input page
}
\description{
get PPP list from an official website
}
\details{
Get PPP list from the Ministry of Finance of China (http://www.cpppc.org:8082/efmisweb/ppp/projectLibrary/toPPPList.do?projName=), to view the listed projects in the PPP library.
}
\examples{
\dontrun{
#scrape the first two pages
getPPPList(1,3)
}
}
\references{
www.cpppc.org
}
