% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBS.R
\name{getWaternomicsData}
\alias{getWaternomicsData}
\title{getWaternomicsData}
\usage{
getWaternomicsData(start, end, url_GDP = NULL, url_wateruse = NULL,
  url_wastewater = NULL)
}
\arguments{
\item{start}{starting year of data wanted}

\item{end}{end year of data wanted, make sure your input end year exists in the NBS website}

\item{url_GDP}{url of GDP sheet}

\item{url_wateruse}{url of Water_use sheet}

\item{url_wastewater}{url of Wastewater}
}
\description{
Get NBS data from google sheet by shared link. Default link is provided by gfer, you can also create your
own google sheet of GDP.
NOTE: The 'link sharing on' of the sheet must be ticked in order to read
}
