% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWQ.R
\name{getWaterQ_MEP_all}
\alias{getWaterQ_MEP_all}
\title{get PPP list from a single page}
\usage{
getWaterQ_MEP_all(year, week, station1, station2, proxy = FALSE)
}
\arguments{
\item{year}{In which year you would like to scrape}

\item{week}{In which week you would like to scrape, can be an array, like 3:5}

\item{station1}{the start station index on the page}

\item{station2}{the end station index on the page}

\item{proxy}{Whether to use proxy, default is FALSE}
}
\description{
get PPP list from a single page
}
\details{
Get monitoring data of different stations from Minitsry of Environmental Protection of China (http://datacenter.mep.gov.cn/report/getCountGraph.do?type=runQianWater). Using this function
you will get data of all the stations. Since the number of stations vary with time, using this function, you have
to make sure that within the period you are scrapping, the number of stations keep consistant.
}
\examples{

\dontrun{
# get data from 1st station to 5th station of the 3rd week of 2016
a <- getWaterQ_MEP_all(2016, 3, 1, 5)

}

}
\references{
http://datacenter.mee.gov.cn/report/getCountGraph.do?type=runQianWater
}
